Separate validator functions from helpers file
