+++ frontend/src/etc/helpers.ts
@@ -26,6 +26,7 @@ import { useNodeDirectiveStore } from "@/stores/nodeDirective";
 import { useObservableTypeStore } from "@/stores/observableType";
 import { useUserStore } from "@/stores/user";
 import { inputTypes } from "./constants/base";
+import { isValidDate, isObject } from "./validators";
 
 export const camelToSnakeCase = (str: string): string =>
   str.replace(/[A-Z]/g, (letter) => `_${letter.toLowerCase()}`);
@@ -114,15 +115,6 @@ export function copyToClipboard(text: string): void | boolean | string | null {
       document.body.removeChild(textarea);
     }
   }
-}
-
-// https://stackoverflow.com/questions/1353684/detecting-an-invalid-date-date-instance-in-javascript
-export function isValidDate(d: unknown): d is Date {
-  return d instanceof Date && !isNaN(d.getTime());
-}
-
-export function isObject(o: unknown): o is Record<string, unknown> {
-  return typeof o === "object" && o !== null;
 }
 
 // https://weblog.west-wind.com/posts/2014/jan/06/javascript-json-date-parsing-and-real-dates
